// src/api/models/role-profile-creation-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RoleProfileCreationModel
 */
export interface RoleProfileCreationModel {
    /**
     * 
     * @type {string}
     * @memberof RoleProfileCreationModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileCreationModel
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileCreationModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof RoleProfileCreationModel
     */
    'isDefault'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileCreationModel
     */
    'businessId'?: number | null;
}