// src/api/models/role-profile-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface RoleProfileEditionModel
 */
export interface RoleProfileEditionModel {
    /**
     * 
     * @type {string}
     * @memberof RoleProfileEditionModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileEditionModel
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RoleProfileEditionModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof RoleProfileEditionModel
     */
    'isDefault'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof RoleProfileEditionModel
     */
    'id'?: number;
    /**
     * 
     * @type {Status}
     * @memberof RoleProfileEditionModel
     */
    'status'?: Status;
}